requirements.txt
Django==4.2.5
asgiref==3.8.1
Deprecated==1.2.14
importlib_metadata==8.0.0
opentelemetry-api==1.26.0
opentelemetry-exporter-prometheus==0.47b0
opentelemetry-instrumentation==0.47b0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
prometheus_client==0.20.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
wrapt==1.16.0
zipp==3.20.0
